Handle date correctly
Now the input is an actual date selector, and the value gets saved to
the database as type datetime. Then it's unmarshalled into integers to
the workout object for passing to the template.

I'm sure there is a "better" way, but this will work for now.

Also some other changes. I haven't commited in a while.

Now it's time to get the delete button working.